+++ scripts/Generate-Message-Sources.ps1
@@ -8,5 +8,5 @@ $rootpath = Join-Path $PSScriptRoot '..' | Resolve-Path
 $specs = Join-Path $rootpath spec fix *xml | Resolve-Path
 
 pushd (Join-Path $rootpath DDTool)
-dotnet run --project DDTool --outputdir $rootpath $specs
+dotnet run --project DDTool --reporoot $rootpath --outputdir $rootpath $specs
 popd