+++ RELEASE_NOTES.md
@@ -59,6 +59,7 @@ What's New
 * #980 - fix: ToJSON returns invalid json when content contains newlines/tabs/etc (Rob-Hague)
 * #309 - fix: obey a SeqReset-GapFill even if it 'replaces' a message that was processed off
          queue in a ResendRequest series (gbirchmeier/oclancy)
+* #979 - improve DDTool: nullable enable, custom name, generate csproj if needed (gbirchmeier)
 
 ### v1.13.1
 * backport #951 to 1.13